즐겨찾기 및 직경 2km 내에 있는 카페 엔드포인트 수정

+++ src/scheduler/updatePlaceData.js
@@ -27,16 +27,29 @@ async function getPlaceDetails(placeId) {
   return response.data.result;
 }
 
-async function getDitsance(origin, destination) {
-  const response = await client.distancematrix({
-    params: {
-      origins: [origin],
-      destinations: [destination],
-      key: process.env.GOOGLE_MAPS_API_KEY,
-    },
-  });
-  return response.data.rows[0].elements[0].distance.value;
+function getDistance(lat1, lon1, lat2, lon2) {
+  const R = 6371; // Earth's radius, km
+  const dLat = deg2rad(lat2 - lat1);
+  const dLon = deg2rad(lon2 - lon1);
+  const a =
+    Math.sin(dLat / 2) * Math.sin(dLat / 2) +
+    Math.cos(deg2rad(lat1)) *
+      Math.cos(deg2rad(lat2)) *
+      Math.sin(dLon / 2) *
+      Math.sin(dLon / 2);
+  const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
+  const distance = R * c; // Distance, km
+  return distance * 1000; // Convert to meters
 }
+
+function deg2rad(deg) {
+  return deg * (Math.PI / 180);
+}
+
+const checkDataByDistance = (lat1, lon1, lat2, lon2, limit = 20) => {
+  const distance = getDistance(lat1, lon1, lat2, lon2);
+  return distance <= limit;
+};
 
 async function getPlacePhoto(photoReference) {
   const response = await client.placePhoto({
@@ -56,30 +69,77 @@ async function main() {
     const allCafes = await locationDao.getAllCafeData();
     let ratesToUpdate = [];
 
-    for (const cafe of allCafes) {
-      const cafeName = cafe.cafe_name;
+    const allTasks = allCafes.map(async (cafe) => {
+      const cafeName = cafe.name;
       const cafeId = cafe.id;
-      const placeId = await getPlaceId(cafeName);
-      const details = await getPlaceDetails(placeId);
 
-      ratesToUpdate.push({
-        cafe_id: cafeId,
-        score: details.rating,
-      });
+      const placeData = await getPlaceId(cafeName);
+      const googleLocation = {
+        lat: placeData.location.lat,
+        lng: placeData.location.lng,
+      };
+      const dbLocation = {
+        lat: cafe.cafe_latitude,
+        lng: cafe.cafe_longitude,
+      };
+
+      if (
+        !checkDataByDistance(
+          googleLocation.lat,
+          googleLocation.lng,
+          dbLocation.lat,
+          dbLocation.lng
+        )
+      ) {
+        return null;
+      }
+
+      const details = await getPlaceDetails(placeData.place_id);
+
+      if (details.rating !== undefined && details.rating !== null) {
+        ratesToUpdate.push({
+          cafe_id: cafeId,
+          score: details.rating,
+        });
+      }
+
+      await locationDao.updateRate(ratesToUpdate);
 
       if (details.photos && details.photos.length > 0) {
         const maxPhotos = Math.min(details.photos.length, 3);
-        for (i = 0; i < maxPhotos; i++) {
+        for (let i = 0; i < maxPhotos; i++) {
+          const imageName = `${cafeName.replace(/ /g, "_")}${i + 1}.jpg`;
+          const savePath = `/Users/khs/Documents/WhichCafe/projectmaterial/cafeimage/${imageName}`;
+
+          const excludedCafes = [
+            "만월경",
+            "에그카페24",
+            "카페일분",
+            "데이롱카페",
+            "커피에반하다",
+            "카페인24",
+          ];
+
+          if (fs.existsSync(savePath) || excludedCafes.includes(cafeName)) {
+            continue;
+          }
+
           const imageData = await getPlacePhoto(
             details.photos[i].photo_reference
           );
-          const imageName = `${cafeName.replace(/ /g, "_")}${i + 1}.jpg`;
-          const savePath = `/Users/khs/Documents/WhichCafe/projectmaterial/cafeimage/${imageName}`;
           fs.writeFileSync(savePath, imageData);
-          result.photoPath = savePath;
+
+          const htmlAttributions = details.html_attributions;
+          await locationDao.updateImgHtml(htmlAttributions, cafeId);
         }
       }
-      await locationDao.updateRate(ratesToUpdate);
+    });
+
+    const results = await Promise.allSettled(allTasks);
+
+    const errors = results.filter((result) => result.status === "rejected");
+    for (const error of errors) {
+      console.error(error.reason);
     }
   } catch (error) {
     console.log(
@@ -88,6 +148,8 @@ async function main() {
   }
 }
 
-schedule.scheduleJob("0 2 1 * *", async function () {
+const scheduledTask = schedule.scheduleJob("0 2 1 * *", async function () {
   await main();
 });
+
+module.exports = { main, scheduledTask };
